commandline/requirements.txt
cryptography==3.4.7
flaky==3.7.0
google-api-python-client==2.1.0
google-auth-httplib2==0.1.0
google-auth==1.28.1
google-cloud-iot==2.0.2
google-cloud-pubsub==1.7.0
paho-mqtt==1.5.1
pyjwt==2.0.1

